requirements.txt
appnope==0.1.0
asgiref==3.2.7
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
brotlipy==0.7.0
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
Django==3.0.3
django-cors-headers==3.2.1
djangorestframework==3.11.0
entrypoints==0.3
idna==2.9
importlib-metadata==1.5.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
nltk==3.4.4
numpy==1.18.1
olefile==0.46
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.4
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
pyzmq==18.1.1
qtconsole==4.7.2
QtPy==1.9.0
requests==2.23.0
scikit-learn==0.22.1
simplegeneric==0.8.1
six==1.14.0
SQLAlchemy==1.3.15
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
wordcloud==1.6.0
zipp==2.2.0
gunicorn==20.0.4
django-heroku==0.3.1
django-q
redis